第一次提交，发现很多问题 写了输入部分需要的两个类，可能还有很多没有写

+++ app/src/main/java/SomePrivateClasses/Rarity.java
@@ -7,9 +7,9 @@ package SomePrivateClasses;
 public class Rarity
 {
     String RarityName;
-    float Probability;
+    double Probability;
 
-    public Rarity(String RarityName,float Probability)
+    public Rarity(String RarityName,double Probability)
     {
         this.RarityName=RarityName;
         this.Probability=Probability;
